requirements.txt
fastapi~=0.115.2
uvicorn
psycopg2-binary
sqlalchemy~=2.0.35

pydantic==2.9.2
PyJWT==2.6.0
passlib~=1.7.4

pydantic-settings~=2.5.2